images not appearing in app
I have created boxes with a grid for the buttons and images (3x3). The images in the images box aren't loading into the app which I think is because the image type of my images is JPG, I have just checked and my computer only supports GIF and PNG image types.